backend/requirements.txt
# backend/requirements.txt
django==4.2.7
django-ninja==1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
dj-database-url==2.1.0
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
statsmodels==0.14.0
empyrical==0.5.5
finnhub-python==2.4.20
gunicorn==21.2.0
django-cors-headers>=3.7.0
requests==2.31.0
python-dateutil==2.8.2
urllib3==2.2.1
whitenoise==6.6.0
debugpy>=1.6.0
django-extensions
django-filter
django-ninja-extra
djangorestframework
uvicorn
cachetools
pytest-django
PyJWT
python-jose 